Fixed bug related to top-level task graph becoming composite task.
1. changed UKCPUTaskCommon_TraverseSubGraphTasks function operation for top-level task graph.
For the hierarchical task, fnCallBack should not be performed for the task itself but for the tasks inside the task. In the past, fnCallBack was performed for the hierarchical task because the code traverses all the tasks of the top-level task graph.
Since callFunctionsInHierarchicalTaskGraph function already performs this hierarchical traversing operation, the callFunctionsInHierarchicalTaskGraph function is used instead of UKTask_TraverseAllTasks function.

2. changed setChannelExitFlags to be occured only if whole task graph is not Dataflow graph.
When one task finishes execution in the last iteration count, the entire program enters the STOPPING state and tries to terminate. This operation is implemented to forcibly terminate a program after a given program execution time is elapsed.
When program terminating, channel exit is also done by setChannelExitFlags. Because of channel termination, tasks that are blocked for waiting channel data receives null data from the channel.
Since forcible channel termination is used only for time-limited example, changed channel termination to be done only for process network example.

Signed-off-by: Dukejung <dowhancool@naver.com>