Implement hierarchical fault tolerant.
CICAlgorithmXMLGenerator.java is mainly changed. Most of them have been referred non-hierarchical fault tolerant implementation.
To make task hierarchical map in CICAlgorithmXMLGenerator, added functions in CICAlgorithmXMLHandler.java.
In regard to DType loop task, it is assumed that when the Dtype loop task has hierarchical structure and is used with multithreading, the same number of cores must be mapped to all internal tasks.
Therefore, implemented max-core assignment function that matches the number of cores of all tasks inside a hierarchical DType loop task to the number of cores in the task to which maximum core is mapped in CICManualDSEPanel.java.

Signed-off-by: Dukejung <dowhancool@naver.com>