[ADD] add api functions to change the period and mapping info
User can change the period of the mapping info of the task.
The period or mapping info can be updated by the control task.
Mapping info can be updated to the different core which is in the same processor.
Also, the only CPU can be used.
Mapping info of the computational or control task can be updated only.

Signed-off-by: JangryulKim <urmydata@snu.ac.kr>